+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
 ## [0.0.4, Unreleased] - 2023-XX-XX
+### Added
+- Generate code coverage report with `make coverage`. 
+  - The coverage report name is `coverage.html`.
+
 ### Changed
 - Dependabot update go.mod in example codes.
+- Introduce `Interval` struct for caluculating interval time.
 
 ### Removed
 - `SpecTest.RecorderHook` is deprecated since apitest package era.